Generated PageView with mix phx.gen.live
Ran the command:

    mix phx.gen.live Analytics PageView page_views hostname:string \
    pathname:string referrer:string

This generates a migration, context, schema, and live views for CRUD
actions.

After this I updated the migration to add null constraints to some
columns that are necessary and removed the updated_at column.

The generator output the following routes, which I added to the router:

    live "/page_views", PageViewLive.Index, :index
    live "/page_views/new", PageViewLive.Index, :new
    live "/page_views/:id/edit", PageViewLive.Index, :edit

    live "/page_views/:id", PageViewLive.Show, :show
    live "/page_views/:id/show/edit", PageViewLive.Show, :edit

Most of the generated LiveView code won't be useful, but we'll prune it
down in subsequent work.